POT VERY OK

+++ week-07/day-1/Dimming_LED_PWM/Dimming_LED_PWM/DIM_LED_POT.h
@@ -1,4 +1,4 @@
 #define DUTY 25
 
 void system_init();
-void duty_time(uint8_t percent);
+void duty_time(int percent);
